requirements.txt
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
requests>=2.32.0
tenacity>=9.0.0
exchange-calendars>=4.5
